geoLocationWebApp/requirements.txt
django-leaflet
geojson
djangorestframework-gis
GDAL